re-arranging the struture of backend
